Move scoring related types to own package
