+++ domain/src/main/kotlin/info/marozzo/hematoma/domain/Tournament.kt
@@ -15,6 +15,7 @@ import arrow.core.raise.zipOrAccumulate
 import arrow.optics.optics
 import info.marozzo.hematoma.domain.errors.Validated
 import info.marozzo.hematoma.domain.errors.ValidationError
+import info.marozzo.hematoma.domain.scoring.Result
 import info.marozzo.hematoma.serializers.PersistentListSerializer
 import info.marozzo.hematoma.serializers.PersistentSetSerializer
 import kotlinx.collections.immutable.*
